felps_trade 1/backend/requirements.txt
# Web
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3

# Ciência de dados / pandas / numpy
numpy==1.26.4
pandas==2.2.2

# Para Cython e builds nativos se necessário
cython==3.0.5

# Para operações assíncronas ou threads (opcional mas avançado)
aiohttp==3.9.9
nest-asyncio==1.5.8

# Outras dependências comuns do seu projeto
python-dotenv==1.0.0
